requirements.txt
# Core dependencies
streamlit==1.41.1
python-dotenv==1.0.1

# LangChain and LangGraph
langchain==0.3.20
langchain-community==0.3.20
langchain-core==0.3.28
langchain-groq==0.2.5
langchain-huggingface==0.1.3
langgraph==0.2.66

# Vector Store and Embeddings
chromadb==0.5.23
sentence-transformers==3.3.1

# Document Processing
pypdf==5.2.0

# HuggingFace and Transformers
huggingface-hub==0.35.3
transformers==4.48.0
torch==2.6.0

# TensorFlow (for embeddings compatibility)
tensorflow==2.20.0
tf-keras==2.20.1

# Protobuf (compatible version)
protobuf==5.29.5

# Additional dependencies
numpy==1.26.4
pydantic==2.10.3
typing-extensions==4.15.0